@bisgaard-itis review: validations
